+++ access-control/src/main/java/com/expediagroup/streamplatform/streamregistry/accesscontrol/domain/AccessType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.streamplatform.streamregistry.core.accesscontrol.domain;
+package com.expediagroup.streamplatform.streamregistry.accesscontrol.domain;
 
 public enum AccessType {