+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/accesscontrol/PermissionEvaluatorImpl.java
@@ -24,8 +24,9 @@ import org.springframework.security.access.PermissionEvaluator;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import com.expediagroup.streamplatform.streamregistry.core.accesscontrol.domain.AccessControlledResource;
-import com.expediagroup.streamplatform.streamregistry.core.accesscontrol.domain.AccessType;
+import com.expediagroup.streamplatform.streamregistry.accesscontrol.AccessControlAuthoriser;
+import com.expediagroup.streamplatform.streamregistry.accesscontrol.domain.AccessControlledResource;
+import com.expediagroup.streamplatform.streamregistry.accesscontrol.domain.AccessType;
 import com.expediagroup.streamplatform.streamregistry.core.accesscontrol.resourcemapper.AccessControlledResourceMapper;
 
 @Component
@@ -59,6 +60,6 @@ public class PermissionEvaluatorImpl implements PermissionEvaluator {
   @Override
   public boolean hasPermission(Authentication authentication, Serializable targetId,
                                String targetType, Object permission) {
-    return true; //TODO handler it later
+    return true;
   }
 }