+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/accesscontrol/resourcemapper/DomainAccessControlledResourceMapper.java
@@ -15,13 +15,13 @@
  */
 package com.expediagroup.streamplatform.streamregistry.core.accesscontrol.resourcemapper;
 
-import static com.expediagroup.streamplatform.streamregistry.core.accesscontrol.domain.AcessControlledResourceFields.DOMAIN_NAME;
+import static com.expediagroup.streamplatform.streamregistry.accesscontrol.domain.AcessControlledResourceFields.DOMAIN_NAME;
 
 import java.util.Map;
 
 import org.springframework.stereotype.Component;
 
-import com.expediagroup.streamplatform.streamregistry.core.accesscontrol.domain.AccessControlledResource;
+import com.expediagroup.streamplatform.streamregistry.accesscontrol.domain.AccessControlledResource;
 import com.expediagroup.streamplatform.streamregistry.model.Domain;
 
 @Component