+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED - 2020-10-14
+### Added
+- Access control for create, update, and delete operations for Consumer, ConsumerBinding, 
+  Domain, Infrastructure, Producer, ProducerBinding, Schema, Stream, StreamBinding, Zone
+
 ## [0.14.4] - 2020-10-05
 ### Added
 - Access control setup to handle who can perform mutations on GraphQL.